fix: cast Reverb port config to integer for Config::integer()
Env vars are always strings, so Config::integer() in HandleInertiaRequests
was failing across feature tests that hit Inertia shared props.

+++ config/broadcasting.php
@@ -39,14 +39,14 @@ return [
             'app_id' => env('REVERB_APP_ID'),
             'options' => [
                 'host' => env('REVERB_HOST'),
-                'port' => env('REVERB_PORT', 443),
+                'port' => (int) env('REVERB_PORT', 443),
                 'scheme' => env('REVERB_SCHEME', 'https'),
                 'useTLS' => env('REVERB_SCHEME', 'https') === 'https',
                 'path' => env('REVERB_SERVER_PATH', ''),
             ],
             'client' => [
                 'host' => env('REVERB_CLIENT_HOST', env('REVERB_HOST', 'localhost')),
-                'port' => env('REVERB_CLIENT_PORT', env('REVERB_PORT', 443)),
+                'port' => (int) env('REVERB_CLIENT_PORT', env('REVERB_PORT', 443)),
                 'scheme' => env('REVERB_CLIENT_SCHEME', env('REVERB_SCHEME', 'https')),
             ],
             'client_options' => [
